put all config files in a dedeicated dir + fixed path in token config module
